Bug fix
Changed the drying time to 20 min

+++ code/ui/src/main/kotlin/uk/ac/gla/student/oip2021/team23/gui/GuiProgress.java
@@ -150,7 +150,7 @@ public class GuiProgress {
     private int currentSequenceRepeated = 0;
     private JDialog dialogUserInput = null;
 
-    private final int timeSignalHeldAcknowledgeTotal = 5000;
+    private final long timeSignalHeldAcknowledgeTotal = 5000;
     private boolean signalWithdrewAcknowledgeSequence = false;
 
     private DetectionThread threadImageDetection = null;
